userAuthentication

+++ controller/expense.js
@@ -9,7 +9,8 @@ exports.AddExpense = async(req,res,next)=>{
     const data = await Expense.create({
         amount:amount,
         description:description,
-        category:category
+        category:category,
+        userId : req.user.id
     })
     res.status(201).json({data:data})
 
@@ -22,7 +23,7 @@ exports.AddExpense = async(req,res,next)=>{
 
 exports.getExpense = async(req,res,next)=>{
     try{ 
-        const expenses = await Expense.findAll();
+        const expenses = await Expense.findAll({where:{userId:req.user.id}});
         res.json({
             expenses
         })
@@ -36,8 +37,18 @@ exports.getExpense = async(req,res,next)=>{
 exports.deleteExpense = async(req,res,next)=>{
     try{
     const id = req.params.expenseId
-    await Expense.destroy({where:{id:id}})
-    res.sendStatus(200)
+    await Expense.destroy({where:{id:id,userId:req.user.id}}).then(noofrowsdeleted=>{
+        console.log(noofrowsdeleted)
+        if(noofrowsdeleted==0){
+            res.status(404).json({sucess:false,message:'Expense doesnt belong to the user'})
+        }
+        else{
+            res.status(200).json({sucess:true,message:'Deleted sucessfully'})
+        }
+    }).catch(err=>{
+        console.log(err)
+    })
+
     }
     catch(err){
         res.json(err)
